+++ ci/e2e.sh
@@ -33,30 +33,6 @@ echo "Add keys for chains"
 echo "-----------------------------------------------------------------------------------------------------------------"
 hermes -c "$CONFIG_PATH" keys add "$CHAIN_A" key_seed_"$CHAIN_A".json
 hermes -c "$CONFIG_PATH" keys add "$CHAIN_B" key_seed_"$CHAIN_B".json
-echo "-----------------------------------------------------------------------------------------------------------------"
-echo "Set the primary peers for clients on each chain                                                                  "
-echo "-----------------------------------------------------------------------------------------------------------------"
-LIGHT_ADD_CHAIN_A="hermes -c $CONFIG_PATH light add tcp://$CHAIN_A:$CHAIN_A_PORT -c $CHAIN_A -s $CHAIN_A_HOME -p -y -f"
-echo "Executing: $LIGHT_ADD_CHAIN_A"
-bash -c "$LIGHT_ADD_CHAIN_A"
-sleep 2
-echo "-----------------------------------------------------------------------------------------------------------------"
-LIGHT_ADD_CHAIN_B="hermes -c $CONFIG_PATH light add tcp://$CHAIN_B:$CHAIN_B_PORT -c $CHAIN_B -s $CHAIN_B_HOME -p -y -f"
-echo "Executing: $LIGHT_ADD_CHAIN_B"
-bash -c "$LIGHT_ADD_CHAIN_B"
-sleep 2
-echo "-----------------------------------------------------------------------------------------------------------------"
-echo "Set the secondary peers for clients on each chain                                                                "
-echo "-----------------------------------------------------------------------------------------------------------------"
-LIGHT_ADD_CHAIN_A_PEER="hermes -c $CONFIG_PATH light add tcp://$CHAIN_A:$CHAIN_A_PORT -c $CHAIN_A -s $CHAIN_A_HOME --peer-id 17D46D8C1576A79203A6733F63B2C9B7235DD559 -y"
-echo "Executing: $LIGHT_ADD_CHAIN_A_PEER"
-bash -c "$LIGHT_ADD_CHAIN_A_PEER"
-sleep 2
-echo "-----------------------------------------------------------------------------------------------------------------"
-LIGHT_ADD_CHAIN_B_PEER="hermes -c $CONFIG_PATH light add tcp://$CHAIN_B:$CHAIN_B_PORT -c $CHAIN_B -s $CHAIN_B_HOME --peer-id A885BB3D3DFF6101188B462466AE926E7A6CD51E -y"
-echo "Executing: $LIGHT_ADD_CHAIN_B_PEER"
-bash -c "$LIGHT_ADD_CHAIN_B_PEER"
-sleep 2
 
 echo "================================================================================================================="
 echo "                                             END-TO-END TESTS                                                    "