+++ relayer-cli/src/commands/tx/channel.rs
@@ -6,9 +6,8 @@ use ibc::ics04_channel::channel::Order;
 use ibc::ics24_host::identifier::{ChainId, ChannelId, ClientId, ConnectionId, PortId};
 use ibc::Height;
 use ibc_relayer::channel::{Channel, ChannelSide};
-use ibc_relayer::config::StoreConfig;
 
-use crate::cli_utils::{ChainHandlePair, SpawnOptions};
+use crate::cli_utils::ChainHandlePair;
 use crate::conclude::Output;
 use crate::error::{Error, Kind};
 use crate::prelude::*;
@@ -17,13 +16,8 @@ macro_rules! tx_chan_cmd {
     ($dbg_string:literal, $func:ident, $self:expr, $chan:expr) => {
         let config = app_config();
 
-        let spawn_options = SpawnOptions::override_store_config(StoreConfig::memory());
-        let chains = match ChainHandlePair::spawn_with(
-            spawn_options,
-            &config,
-            &$self.src_chain_id,
-            &$self.dst_chain_id,
-        ) {
+        let chains = match ChainHandlePair::spawn(&config, &$self.src_chain_id, &$self.dst_chain_id)
+        {
             Ok(chains) => chains,
             Err(e) => return Output::error(format!("{}", e)).exit(),
         };