+++ relayer/src/chain/cosmos.rs
@@ -62,7 +62,7 @@ use crate::config::ChainConfig;
 use crate::error::{Error, Kind};
 use crate::event::monitor::{EventBatch, EventMonitor};
 use crate::keyring::store::{KeyEntry, KeyRing, KeyRingOperations, StoreBackend};
-use crate::light_client::tendermint::LightClient as TMLightClient;
+use crate::light_client::tendermint::LightClient as TmLightClient;
 use crate::light_client::LightClient;
 
 use super::Chain;
@@ -78,6 +78,7 @@ const DEFAULT_GAS_FEE_AMOUNT: u64 = 1000;
 pub struct CosmosSdkChain {
     config: ChainConfig,
     rpc_client: HttpClient,
+    grpc_addr: Uri,
     rt: Arc<TokioRuntime>,
     keybase: KeyRing,
 }
@@ -87,13 +88,11 @@ impl CosmosSdkChain {
     pub fn unbonding_period(&self) -> Result<Duration, Error> {
         crate::time!("unbonding_period");
 
-        // TODO - generalize this
-        let grpc_addr =
-            Uri::from_str(&self.config().grpc_addr).map_err(|e| Kind::Grpc.context(e))?;
-
         let mut client = self
             .block_on(
-                ibc_proto::cosmos::staking::v1beta1::query_client::QueryClient::connect(grpc_addr),
+                ibc_proto::cosmos::staking::v1beta1::query_client::QueryClient::connect(
+                    self.grpc_addr.clone(),
+                ),
             )
             .map_err(|e| Kind::Grpc.context(e))?;
 
@@ -321,29 +320,35 @@ impl Chain for CosmosSdkChain {
             .map_err(|e| Kind::Rpc(config.rpc_addr.clone()).context(e))?;
 
         // Initialize key store and load key
-        let key_store = KeyRing::init(StoreBackend::Test, config.clone())
+        let keybase = KeyRing::init(StoreBackend::Test, config.clone())
             .map_err(|e| Kind::KeyBase.context(e))?;
+
+        let grpc_addr =
+            Uri::from_str(&config.grpc_addr.to_string()).map_err(|e| Kind::Grpc.context(e))?;
 
         Ok(Self {
             rt,
             config,
-            keybase: key_store,
+            keybase,
             rpc_client,
+            grpc_addr,
         })
     }
 
-    // TODO use a simpler approach to create the light client
-    #[allow(clippy::type_complexity)]
-    fn init_light_client(
-        &self,
-    ) -> Result<(Box<dyn LightClient<Self>>, Option<thread::JoinHandle<()>>), Error> {
+    fn init_light_client(&self) -> Result<Box<dyn LightClient<Self>>, Error> {
+        use tendermint_light_client::types::PeerId;
+
         crate::time!("init_light_client");
 
-        let (lc, supervisor) = TMLightClient::from_config(&self.config, true)?;
+        let peer_id: PeerId = self
+            .rt
+            .block_on(self.rpc_client.status())
+            .map(|s| s.node_info.id)
+            .map_err(|e| Kind::Rpc(self.config.rpc_addr.clone()).context(e))?;
 
-        let supervisor_thread = thread::spawn(move || supervisor.run().unwrap());
+        let light_client = TmLightClient::from_config(&self.config, peer_id)?;
 
-        Ok((Box::new(lc), Some(supervisor_thread)))
+        Ok(Box::new(light_client))
     }
 
     fn init_event_monitor(
@@ -358,8 +363,11 @@ impl Chain for CosmosSdkChain {
     > {
         crate::time!("init_event_monitor");
 
-        let (mut event_monitor, event_receiver) =
-            EventMonitor::new(self.config.id.clone(), self.config.rpc_addr.clone(), rt)?;
+        let (mut event_monitor, event_receiver) = EventMonitor::new(
+            self.config.id.clone(),
+            self.config.websocket_addr.clone(),
+            rt,
+        )?;
 
         event_monitor.subscribe().unwrap();
         let monitor_thread = thread::spawn(move || event_monitor.run());
@@ -455,7 +463,7 @@ impl Chain for CosmosSdkChain {
 
         if status.sync_info.catching_up {
             fail!(
-                Kind::LightClientSupervisor(self.config.id.clone()),
+                Kind::LightClient(self.config.rpc_addr.to_string()),
                 "node at {} running chain {} not caught up",
                 self.config().rpc_addr,
                 self.config().id,
@@ -471,12 +479,11 @@ impl Chain for CosmosSdkChain {
     fn query_clients(&self, request: QueryClientStatesRequest) -> Result<Vec<ClientId>, Error> {
         crate::time!("query_chain_clients");
 
-        let grpc_addr =
-            Uri::from_str(&self.config().grpc_addr).map_err(|e| Kind::Grpc.context(e))?;
-
         let mut client = self
             .block_on(
-                ibc_proto::ibc::core::client::v1::query_client::QueryClient::connect(grpc_addr),
+                ibc_proto::ibc::core::client::v1::query_client::QueryClient::connect(
+                    self.grpc_addr.clone(),
+                ),
             )
             .map_err(|e| Kind::Grpc.context(e))?;
 
@@ -522,13 +529,10 @@ impl Chain for CosmosSdkChain {
     ) -> Result<(Self::ClientState, MerkleProof), Error> {
         crate::time!("query_upgraded_client_state");
 
-        let grpc_address =
-            Uri::from_str(&self.config.grpc_addr).map_err(|e| Kind::Grpc.context(e))?;
-
         let mut client = self
             .block_on(
                 ibc_proto::cosmos::upgrade::v1beta1::query_client::QueryClient::connect(
-                    grpc_address,
+                    self.grpc_addr.clone(),
                 ),
             )
             .map_err(|e| Kind::Grpc.context(e))?;
@@ -573,13 +577,10 @@ impl Chain for CosmosSdkChain {
         let tm_height =
             Height::try_from(height.revision_height).map_err(|e| Kind::InvalidHeight.context(e))?;
 
-        let grpc_address =
-            Uri::from_str(&self.config.grpc_addr).map_err(|e| Kind::Grpc.context(e))?;
-
         let mut client = self
             .block_on(
                 ibc_proto::cosmos::upgrade::v1beta1::query_client::QueryClient::connect(
-                    grpc_address,
+                    self.grpc_addr.clone(),
                 ),
             )
             .map_err(|e| Kind::Grpc.context(e))?;
@@ -622,11 +623,11 @@ impl Chain for CosmosSdkChain {
     ) -> Result<Vec<ConnectionId>, Error> {
         crate::time!("query_connections");
 
-        let grpc_addr =
-            Uri::from_str(&self.config().grpc_addr).map_err(|e| Kind::Grpc.context(e))?;
         let mut client = self
             .block_on(
-                ibc_proto::ibc::core::connection::v1::query_client::QueryClient::connect(grpc_addr),
+                ibc_proto::ibc::core::connection::v1::query_client::QueryClient::connect(
+                    self.grpc_addr.clone(),
+                ),
             )
             .map_err(|e| Kind::Grpc.context(e))?;
 
@@ -655,11 +656,11 @@ impl Chain for CosmosSdkChain {
     ) -> Result<Vec<ConnectionId>, Error> {
         crate::time!("query_connections");
 
-        let grpc_addr =
-            Uri::from_str(&self.config().grpc_addr).map_err(|e| Kind::Grpc.context(e))?;
         let mut client = self
             .block_on(
-                ibc_proto::ibc::core::connection::v1::query_client::QueryClient::connect(grpc_addr),
+                ibc_proto::ibc::core::connection::v1::query_client::QueryClient::connect(
+                    self.grpc_addr.clone(),
+                ),
             )
             .map_err(|e| Kind::Grpc.context(e))?;
 
@@ -698,11 +699,11 @@ impl Chain for CosmosSdkChain {
     ) -> Result<Vec<ChannelId>, Error> {
         crate::time!("query_connection_channels");
 
-        let grpc_addr =
-            Uri::from_str(&self.config().grpc_addr).map_err(|e| Kind::Grpc.context(e))?;
         let mut client = self
             .block_on(
-                ibc_proto::ibc::core::channel::v1::query_client::QueryClient::connect(grpc_addr),
+                ibc_proto::ibc::core::channel::v1::query_client::QueryClient::connect(
+                    self.grpc_addr.clone(),
+                ),
             )
             .map_err(|e| Kind::Grpc.context(e))?;
 
@@ -728,11 +729,11 @@ impl Chain for CosmosSdkChain {
     fn query_channels(&self, request: QueryChannelsRequest) -> Result<Vec<ChannelId>, Error> {
         crate::time!("query_connections");
 
-        let grpc_addr =
-            Uri::from_str(&self.config().grpc_addr).map_err(|e| Kind::Grpc.context(e))?;
         let mut client = self
             .block_on(
-                ibc_proto::ibc::core::channel::v1::query_client::QueryClient::connect(grpc_addr),
+                ibc_proto::ibc::core::channel::v1::query_client::QueryClient::connect(
+                    self.grpc_addr.clone(),
+                ),
             )
             .map_err(|e| Kind::Grpc.context(e))?;
 
@@ -777,11 +778,11 @@ impl Chain for CosmosSdkChain {
     ) -> Result<(Vec<PacketState>, ICSHeight), Error> {
         crate::time!("query_packet_commitments");
 
-        let grpc_addr =
-            Uri::from_str(&self.config().grpc_addr).map_err(|e| Kind::Grpc.context(e))?;
         let mut client = self
             .block_on(
-                ibc_proto::ibc::core::channel::v1::query_client::QueryClient::connect(grpc_addr),
+                ibc_proto::ibc::core::channel::v1::query_client::QueryClient::connect(
+                    self.grpc_addr.clone(),
+                ),
             )
             .map_err(|e| Kind::Grpc.context(e))?;
 
@@ -810,11 +811,11 @@ impl Chain for CosmosSdkChain {
     ) -> Result<Vec<u64>, Error> {
         crate::time!("query_unreceived_packets");
 
-        let grpc_addr =
-            Uri::from_str(&self.config().grpc_addr).map_err(|e| Kind::Grpc.context(e))?;
         let mut client = self
             .block_on(
-                ibc_proto::ibc::core::channel::v1::query_client::QueryClient::connect(grpc_addr),
+                ibc_proto::ibc::core::channel::v1::query_client::QueryClient::connect(
+                    self.grpc_addr.clone(),
+                ),
             )
             .map_err(|e| Kind::Grpc.context(e))?;
 
@@ -836,11 +837,11 @@ impl Chain for CosmosSdkChain {
     ) -> Result<(Vec<PacketState>, ICSHeight), Error> {
         crate::time!("query_packet_acknowledgements");
 
-        let grpc_addr =
-            Uri::from_str(&self.config().grpc_addr).map_err(|e| Kind::Grpc.context(e))?;
         let mut client = self
             .block_on(
-                ibc_proto::ibc::core::channel::v1::query_client::QueryClient::connect(grpc_addr),
+                ibc_proto::ibc::core::channel::v1::query_client::QueryClient::connect(
+                    self.grpc_addr.clone(),
+                ),
             )
             .map_err(|e| Kind::Grpc.context(e))?;
 
@@ -869,11 +870,11 @@ impl Chain for CosmosSdkChain {
     ) -> Result<Vec<u64>, Error> {
         crate::time!("query_unreceived_acknowledgements");
 
-        let grpc_addr =
-            Uri::from_str(&self.config().grpc_addr).map_err(|e| Kind::Grpc.context(e))?;
         let mut client = self
             .block_on(
-                ibc_proto::ibc::core::channel::v1::query_client::QueryClient::connect(grpc_addr),
+                ibc_proto::ibc::core::channel::v1::query_client::QueryClient::connect(
+                    self.grpc_addr.clone(),
+                ),
             )
             .map_err(|e| Kind::Grpc.context(e))?;
 
@@ -894,11 +895,11 @@ impl Chain for CosmosSdkChain {
     ) -> Result<Sequence, Error> {
         crate::time!("query_next_sequence_receive");
 
-        let grpc_addr =
-            Uri::from_str(&self.config().grpc_addr).map_err(|e| Kind::Grpc.context(e))?;
         let mut client = self
             .block_on(
-                ibc_proto::ibc::core::channel::v1::query_client::QueryClient::connect(grpc_addr),
+                ibc_proto::ibc::core::channel::v1::query_client::QueryClient::connect(
+                    self.grpc_addr.clone(),
+                ),
             )
             .map_err(|e| Kind::Grpc.context(e))?;
 
@@ -1290,11 +1291,11 @@ async fn broadcast_tx_commit(
 
 /// Uses the GRPC client to retrieve the account sequence
 async fn query_account(chain: &CosmosSdkChain, address: String) -> Result<BaseAccount, Error> {
-    let grpc_addr = Uri::from_str(&chain.config().grpc_addr).map_err(|e| Kind::Grpc.context(e))?;
-    let mut client =
-        ibc_proto::cosmos::auth::v1beta1::query_client::QueryClient::connect(grpc_addr)
-            .await
-            .map_err(|e| Kind::Grpc.context(e))?;
+    let mut client = ibc_proto::cosmos::auth::v1beta1::query_client::QueryClient::connect(
+        chain.grpc_addr.clone(),
+    )
+    .await
+    .map_err(|e| Kind::Grpc.context(e))?;
 
     let request = tonic::Request::new(QueryAccountRequest { address });